requirements.txt
requests>=2.31.0
openai>=1.12.0
azure-ai-inference==1.0.0b9
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
rich>=13.7.0
pytz>=2024.1